+++ pytest_inmanta/plugin.py
@@ -19,30 +19,38 @@ import glob
 import importlib
 import json
 import logging
+import math
 import os
 import shutil
 import sys
 import tempfile
+import typing
 import warnings
 from collections import defaultdict
 from distutils import dir_util
 from pathlib import Path
+from textwrap import dedent
 from types import FunctionType, ModuleType
-from typing import Dict, Iterator, List, Optional, Tuple, Union
 
 import pytest
 import yaml
+from pytest import CaptureFixture
 from tornado import ioloop
 
+import inmanta.ast
 from inmanta import compiler, config, const, module, protocol
 from inmanta.agent import cache, handler
 from inmanta.agent import io as agent_io
 from inmanta.agent.handler import HandlerContext, ResourceHandler
-from inmanta.data import LogLine
+from inmanta.data import LogLine, ResourceIdStr
+from inmanta.data.model import AttributeStateChange
 from inmanta.execute.proxy import DynamicProxy
-from inmanta.export import Exporter, cfg_env
+from inmanta.export import Exporter, ResourceDict, cfg_env
 from inmanta.protocol import json_encode
 from inmanta.resources import Resource
+
+if typing.TYPE_CHECKING:
+    from inmanta.agent.io.local import IOBase
 
 from .handler import DATA
 
@@ -58,7 +66,7 @@ option_to_env = {
 }
 
 
-def pytest_addoption(parser):
+def pytest_addoption(parser) -> None:
     group = parser.getgroup("inmanta", "inmanta module testing plugin")
     group.addoption(
         "--venv",
@@ -102,7 +110,7 @@ def pytest_addoption(parser):
     )
 
 
-def get_opt_or_env_or(config, key, default):
+def get_opt_or_env_or(config, key: str, default: str) -> str:
     if config.getoption(key):
         return config.getoption(key)
     if option_to_env[key] in os.environ:
@@ -110,7 +118,7 @@ def get_opt_or_env_or(config, key, default):
     return default
 
 
-def get_module_info():
+def get_module_info() -> typing.Tuple[str, str]:
     curdir = CURDIR
     # Make sure that we are executed in a module
     dir_path = curdir.split(os.path.sep)
@@ -134,13 +142,17 @@ def get_module_info():
 
 
 @pytest.fixture()
-def inmanta_plugins(project):
+def inmanta_plugins(
+    project: "Project",
+) -> typing.Iterator["InmantaPluginsImportLoader"]:
     importer: InmantaPluginsImporter = InmantaPluginsImporter(project)
     yield importer.loader
 
 
 @pytest.fixture()
-def project(project_shared, capsys):
+def project(
+    project_shared: "Project", capsys: CaptureFixture
+) -> typing.Iterator["Project"]:
     DATA.clear()
     project_shared.clean()
     project_shared.init(capsys)
@@ -149,7 +161,9 @@ def project(project_shared, capsys):
 
 
 @pytest.fixture()
-def project_no_plugins(project_shared_no_plugins, capsys):
+def project_no_plugins(
+    project_shared_no_plugins, capsys: CaptureFixture
+) -> typing.Iterator["Project"]:
     warnings.warn(
         DeprecationWarning(
             "The project_no_plugins fixture is deprecated in favor of the %s environment variable."
@@ -173,7 +187,9 @@ def get_module_data(filename: str) -> str:
 
 
 @pytest.fixture(scope="session")
-def project_shared(project_factory):
+def project_shared(
+    project_factory: typing.Callable[[typing.Dict[str, typing.Any]], "Project"]
+) -> "Project":
     """
     A test fixture that creates a new inmanta project with the current module in. The returned object can be used
     to add files to the unittest module, compile a model and access the results, stdout and stderr.
@@ -183,7 +199,9 @@ def project_shared(project_factory):
 
 # Temporary workaround for plugins loading multiple times (inmanta/pytest-inmanta#49)
 @pytest.fixture(scope="session")
-def project_shared_no_plugins(project_factory):
+def project_shared_no_plugins(
+    project_factory: typing.Callable[[typing.Dict[str, typing.Any]], "Project"]
+) -> "Project":
     """
     A test fixture that creates a new inmanta project with the current module in. The returned object can be used
     to add files to the unittest module, compile a model and access the results, stdout and stderr.
@@ -193,7 +211,7 @@ def project_shared_no_plugins(project_factory):
 
 
 @pytest.fixture(scope="session")
-def project_factory(request):
+def project_factory(request: pytest.FixtureRequest) -> typing.Callable[[], "Project"]:
     """
     A factory that constructs a single Project.
     """
@@ -256,7 +274,7 @@ install_mode: %(install_mode)s
         )
 
     def create_project(**kwargs):
-        extended_kwargs: Dict[str, object] = {
+        extended_kwargs: typing.Dict[str, object] = {
             "load_plugins": not get_opt_or_env_or(
                 request.config, "inm_no_load_plugins", False
             ),
@@ -317,9 +335,9 @@ class InmantaPluginsImportLoader:
         self._importer: InmantaPluginsImporter = importer
 
     def __getattr__(self, name: str):
-        submodules: Optional[Dict[str, ModuleType]] = self._importer.get_submodules(
-            name
-        )
+        submodules: typing.Optional[
+            typing.Dict[str, ModuleType]
+        ] = self._importer.get_submodules(name)
         fq_mod_name: str = f"inmanta_plugins.{name}"
         if submodules is None or fq_mod_name not in submodules:
             raise AttributeError("No inmanta module named %s" % name)
@@ -331,14 +349,16 @@ class InmantaPluginsImporter:
         self.project: Project = project
         self.loader: InmantaPluginsImportLoader = InmantaPluginsImportLoader(self)
 
-    def get_submodules(self, module_name: str) -> Optional[Dict[str, ModuleType]]:
+    def get_submodules(
+        self, module_name: str
+    ) -> typing.Optional[typing.Dict[str, ModuleType]]:
         inmanta_project: module.Project = module.Project.get()
         if not inmanta_project.loaded:
             raise Exception(
                 "Dynamically importing from inmanta_plugins requires a loaded inmanta.module.Project. Make sure to use the"
                 " project fixture."
             )
-        modules: Dict[str, module.Module] = inmanta_project.get_modules()
+        modules: typing.Dict[str, module.Module] = inmanta_project.get_modules()
         if module_name not in modules:
             return None
         result = {}
@@ -350,7 +370,7 @@ class InmantaPluginsImporter:
     # TODO: this method duplicates inmanta.module.Module.get_plugin_files (2020.4), see #76
     def get_plugin_files_for_module(
         self, mod: module.Module
-    ) -> Iterator[Tuple[str, str]]:
+    ) -> typing.Iterator[typing.Tuple[str, str]]:
         """
         Returns a tuple (absolute_path, fq_mod_name) of all python files in this module.
         """
@@ -382,34 +402,36 @@ class Project:
     environment variable. Repositories are separated with spaces.
     """
 
-    def __init__(self, project_dir, load_plugins: bool = True):
+    def __init__(self, project_dir: str, load_plugins: bool = True) -> None:
         self._test_project_dir = project_dir
-        self._stdout = None
-        self._stderr = None
-        self.types = None
-        self.version = None
-        self.resources = {}
-        self._root_scope = {}
-        self._exporter = None
-        self._blobs = {}
-        self._facts = defaultdict(dict)
+        self._stdout: typing.Optional[str] = None
+        self._stderr: typing.Optional[str] = None
+        self.types: typing.Optional[typing.Dict[str, inmanta.ast.Type]] = None
+        self.version: typing.Optional[int] = None
+        self.resources: ResourceDict = {}
+        self._root_scope: typing.Optional[inmanta.ast.Namespace] = None
+        self._exporter: typing.Optional[Exporter] = None
+        self._blobs: typing.Dict[str, bytes] = {}
+        self._facts: typing.Dict[
+            ResourceIdStr, typing.Dict[str, typing.Any]
+        ] = defaultdict(dict)
         self._load()
-        self._plugins: Optional[Dict[str, object]] = (
+        self._plugins: typing.Optional[typing.Dict[str, FunctionType]] = (
             self._load_plugins() if load_plugins else None
         )
-        self._capsys = None
-        self.ctx = None
-        self._handlers = set()
+        self._capsys: typing.Optional[CaptureFixture] = None
+        self.ctx: typing.Optional[HandlerContext] = None
+        self._handlers: typing.Set[ResourceHandler] = set()
         config.Config.load_config()
 
-    def init(self, capsys):
+    def init(self, capsys: CaptureFixture) -> None:
         self._stdout = None
         self._stderr = None
         self._capsys = capsys
         self.types = None
         self.version = None
         self.resources = {}
-        self._root_scope = {}
+        self._root_scope = None
         self._exporter = None
         self._blobs = {}
         self._facts = defaultdict(dict)
@@ -417,24 +439,27 @@ class Project:
         self._handlers = set()
         config.Config.load_config()
 
-    def add_blob(self, key, content, allow_overwrite=True):
+    def add_blob(self, key: str, content: bytes, allow_overwrite: bool = True) -> None:
         """
         Add a blob identified with the hash of the content as key
         """
+        if isinstance(content, str):
+            warnings.warn("received a string, but expect bytes", DeprecationWarning)
+            content = content.encode("utf-8")
         if key in self._blobs and not allow_overwrite:
             raise Exception("Key %s already stored in blobs" % key)
         self._blobs[key] = content
 
-    def stat_blob(self, key):
+    def stat_blob(self, key: str) -> bool:
         return key in self._blobs
 
-    def get_blob(self, key):
+    def get_blob(self, key: str) -> bytes:
         return self._blobs[key]
 
-    def add_fact(self, resource_id, name, value):
+    def add_fact(self, resource_id: ResourceIdStr, name: str, value: object) -> None:
         self._facts[resource_id][name] = value
 
-    def get_handler(self, resource, run_as_root):
+    def get_handler(self, resource: Resource, run_as_root: bool) -> ResourceHandler:
         # TODO: if user is root, do not use remoting
         c = cache.AgentCache()
         if run_as_root:
@@ -444,22 +469,24 @@ class Project:
 
         c.open_version(resource.id.version)
         try:
-            p = handler.Commander.get_provider(c, agent, resource)
+            p = handler.Commander.get_provider(c, agent, resource)  # typing: ignore
             p.set_cache(c)
-            p.get_file = lambda x: self.get_blob(x)
-            p.stat_file = lambda x: self.stat_blob(x)
-            p.upload_file = lambda x, y: self.add_blob(x, y)
-            p.run_sync = ioloop.IOLoop.current().run_sync
+            p.get_file = lambda x: self.get_blob(x)  # typing: ignore
+            p.stat_file = lambda x: self.stat_blob(x)  # typing: ignore
+            p.upload_file = lambda x, y: self.add_blob(x, y)  # typing: ignore
+            p.run_sync = ioloop.IOLoop.current().run_sync  # typing: ignore
             self._handlers.add(p)
             return p
         except Exception as e:
             raise e
 
-    def finalize_context(self, ctx: handler.HandlerContext):
+    def finalize_context(self, ctx: handler.HandlerContext) -> None:
         # ensure logs can be serialized
         json_encode({"message": ctx.logs})
 
-    def get_resource(self, resource_type: str, **filter_args: dict):
+    def get_resource(
+        self, resource_type: str, **filter_args: typing.Dict[str, object]
+    ) -> typing.Optional[Resource]:
         """
         Get a resource of the given type and given filter on the resource attributes. If multiple resource match, the
         first one is returned. If none match, None is returned.
@@ -467,7 +494,7 @@ class Project:
         :param resource_type: The exact type used in the model (no super types)
         """
 
-        def apply_filter(resource):
+        def apply_filter(resource: Resource) -> bool:
             for arg, value in filter_args.items():
                 if not hasattr(resource, arg):
                     return False
@@ -489,7 +516,9 @@ class Project:
 
         return None
 
-    def deploy(self, resource, dry_run=False, run_as_root=False):
+    def deploy(
+        self, resource: Resource, dry_run: bool = False, run_as_root: bool = False
+    ) -> HandlerContext:
         """
         Deploy the given resource with a handler
         """
@@ -506,16 +535,28 @@ class Project:
         self.finalize_handler(h)
         return ctx
 
-    def dryrun(self, resource, run_as_root=False):
+    def dryrun(self, resource: Resource, run_as_root: bool = False) -> HandlerContext:
         return self.deploy(resource, True, run_as_root)
 
     def deploy_resource(
         self,
         resource_type: str,
-        status=const.ResourceState.deployed,
-        run_as_root=False,
-        **filter_args: dict,
-    ):
+        status: const.ResourceState = const.ResourceState.deployed,
+        run_as_root: bool = False,
+        change: const.Change = None,
+        **filter_args: typing.Dict[str, object],
+    ) -> Resource:
+        """
+        Deploy a resource of the given type, that matches the filter and assert the outcome
+
+        :param resource_type: the type of resource to deploy
+        :param filter_args: a set of kwargs, the resource must have all matching attributes set to the given values
+        :param run_as_root: run the handler as root or not
+        :param status: the expected status of the deployment
+        :param change: the expected change performed by the handler
+
+        :return: the resource
+        """
         res = self.get_resource(resource_type, **filter_args)
         assert res is not None, "No resource found of given type and filter args"
 
@@ -537,16 +578,18 @@ class Project:
                     print("Traceback:\n", log._data["kwargs"]["traceback"])
 
         assert ctx.status == status
+        if change is not None:
+            assert ctx._change == change
         self.finalize_context(ctx)
         return res
 
     def dryrun_resource(
         self,
         resource_type: str,
-        status=const.ResourceState.dry,
-        run_as_root=False,
-        **filter_args: dict,
-    ):
+        status: const.ResourceState = const.ResourceState.dry,
+        run_as_root: bool = False,
+        **filter_args: typing.Dict[str, object],
+    ) -> typing.Dict[str, AttributeStateChange]:
         """
         Run a dryrun for a specific resource.
 
@@ -564,7 +607,7 @@ class Project:
         assert ctx.status == status
         return ctx.changes
 
-    def io(self, run_as_root=False):
+    def io(self, run_as_root: bool = False) -> "IOBase":
         version = 1
         if run_as_root:
             ret = agent_io.get_io(None, "ssh://root@localhost", version)
@@ -572,7 +615,7 @@ class Project:
             ret = agent_io.get_io(None, "local:", version)
         return ret
 
-    def create_module(self, name, initcf="", initpy=""):
+    def create_module(self, name: str, initcf: str = "", initpy: str = "") -> None:
         module_dir = os.path.join(self._test_project_dir, "libs", name)
         os.mkdir(module_dir)
         os.mkdir(os.path.join(module_dir, "model"))
@@ -605,13 +648,34 @@ license: Test License
         module.Project.set(test_project)
         test_project.load()
 
-    def compile(self, main, export=False):
+    def compile(self, main: str, export: bool = False, no_dedent: bool = True) -> None:
         """
         Compile the configuration model in main. This method will load all required modules.
+
+        :param main: The model to compile
+        :param export: Whether the model should be exported after the compile
+        :param no_dedent: Don't remove additional indentation in the model
         """
+        # Dedent the input format
+        model = dedent(main.strip("\n")) if not no_dedent else main
+
         # write main.cf
         with open(os.path.join(self._test_project_dir, "main.cf"), "w+") as fd:
-            fd.write(main)
+            fd.write(model)
+
+        # logging model with line numbers
+        def enumerate_model(model: str):
+            lines = model.split("\n")
+            leading_zeros = math.floor(math.log(len(lines), 10)) + 1
+            line_numbers_model = "\n".join(
+                [
+                    f"{str(number).zfill(leading_zeros)}    {line}"
+                    for number, line in enumerate(lines, start=1)
+                ]
+            )
+            return line_numbers_model
+
+        LOGGER.debug(f"Compiling model:\n{enumerate_model(model)}")
 
         # compile the model
         test_project = module.Project(self._test_project_dir)
@@ -644,10 +708,12 @@ license: Test License
         self._stdout = captured.out
         self._stderr = captured.err
 
-    def deploy_latest_version(self, full_deploy=False):
+    def deploy_latest_version(self, full_deploy: bool = False) -> None:
         """Release and push the latest version to the server (uses the current configuration, either with a fixture or
         set by the test.
         """
+        if self.version is None:
+            raise Exception("Run project.compile first")
         conn = protocol.SyncClient("compiler")
         LOGGER.info("Triggering deploy for version %d" % self.version)
         tid = cfg_env.get()
@@ -656,24 +722,24 @@ license: Test License
         )
         conn.release_version(tid, self.version, True, agent_trigger_method)
 
-    def get_last_context(self) -> Optional[HandlerContext]:
+    def get_last_context(self) -> typing.Optional[HandlerContext]:
         return self.ctx
 
-    def get_last_logs(self) -> Optional[List[LogLine]]:
+    def get_last_logs(self) -> typing.Optional[typing.List[LogLine]]:
         if self.ctx:
             return self.ctx.logs
         return None
 
-    def get_stdout(self):
+    def get_stdout(self) -> typing.Optional[str]:
         return self._stdout
 
-    def get_stderr(self):
+    def get_stderr(self) -> typing.Optional[str]:
         return self._stderr
 
-    def get_root_scope(self):
+    def get_root_scope(self) -> typing.Optional[inmanta.ast.Namespace]:
         return self._root_scope
 
-    def add_mock_file(self, subdir, name, content):
+    def add_mock_file(self, subdir: str, name: str, content: str) -> None:
         """
         This method can be used to register mock templates or files in the virtual "unittest" module.
         """
@@ -684,11 +750,11 @@ license: Test License
         with open(os.path.join(dir_name, name), "w+") as fd:
             fd.write(content)
 
-    def _load_plugins(self) -> Dict[str, FunctionType]:
+    def _load_plugins(self) -> typing.Dict[str, FunctionType]:
         _, module_name = get_module_info()
-        submodules: Optional[Dict[str, ModuleType]] = InmantaPluginsImporter(
-            self
-        ).get_submodules(module_name)
+        submodules: typing.Optional[
+            typing.Dict[str, ModuleType]
+        ] = InmantaPluginsImporter(self).get_submodules(module_name)
         return (
             {}
             if submodules is None
@@ -700,7 +766,7 @@ license: Test License
             }
         )
 
-    def get_plugin_function(self, function_name):
+    def get_plugin_function(self, function_name: str) -> FunctionType:
         if self._plugins is None:
             raise Exception(
                 "Plugins not loaded, perhaps you should use the `project` fixture or"
@@ -710,7 +776,7 @@ license: Test License
             raise Exception(f"Plugin function with name {function_name} not found")
         return self._plugins[function_name]
 
-    def get_plugins(self):
+    def get_plugins(self) -> typing.Dict[str, FunctionType]:
         if self._plugins is None:
             raise Exception(
                 "Plugins not loaded, perhaps you should use the `project` fixture or"
@@ -718,7 +784,9 @@ license: Test License
             )
         return dict(self._plugins)
 
-    def get_instances(self, fortype: str = "std::Entity"):
+    def get_instances(self, fortype: str = "std::Entity") -> typing.List[DynamicProxy]:
+        if self.types is None:
+            raise Exception("No compile has been done")
         if fortype not in self.types:
             raise Exception(f"No entities of type {fortype} found in the model")
 
@@ -736,14 +804,14 @@ license: Test License
 
     def unittest_resource_get(
         self, name: str
-    ) -> Dict[str, Union[str, bool, float, int]]:
+    ) -> typing.Dict[str, typing.Union[str, bool, float, int]]:
         """
         Get the state of the unittest resource
         """
         return DATA[name]
 
     def unittest_resource_set(
-        self, name: str, **kwargs: Union[str, bool, float, int]
+        self, name: str, **kwargs: typing.Union[str, bool, float, int]
     ) -> None:
         """
         Change a value of the unittest resource
@@ -767,6 +835,6 @@ license: Test License
     def finalize_handler(self, handler: ResourceHandler) -> None:
         handler.cache.close()
 
-    def finalize_all_handlers(self):
+    def finalize_all_handlers(self) -> None:
         for handler_instance in self._handlers:
             self.finalize_handler(handler_instance)